feat: PerturbationsFilterSortStore separated into two files + created Props file for ControlPerturbationTable page
