+++ src/pages/control-perturbations-table/ControlPerturbationTableProps.ts
@@ -1,0 +1,1 @@
+export type ControlPerturbationTableProps = {};
