+++ src/stores/ControlPerturbationsTable/PerturbationsFilterSortStore/usePerturbationsFilterSortStore.ts
@@ -1,25 +1,6 @@
 import { create } from 'zustand';
-import type { PertTableSort, PertVariableFilterStatus } from '../../types';
-
-type PerturbationFiltersSortState = {
-  pageNumber: number;
-  perturbationVariables: Record<string, PertVariableFilterStatus>;
-  minRobustness: number | undefined;
-  maxSize: number | undefined;
-  minNumberOfInterpretations: number | undefined;
-  primarySort: PertTableSort | undefined;
-  secondarySort: PertTableSort | undefined;
-  setPageNumber: (page: number) => void;
-  setPerturbationVariables: (
-    vars: Record<string, PertVariableFilterStatus>
-  ) => void;
-  setMinRobustness: (value: number | undefined) => void;
-  setMaxSize: (value: number | undefined) => void;
-  setMinNumberOfInterpretations: (value: number | undefined) => void;
-  setPrimarySort: (sort: PertTableSort | undefined) => void;
-  setSecondarySort: (sort: PertTableSort | undefined) => void;
-  clear: () => void;
-};
+import type { PertTableSort, PertVariableFilterStatus } from '../../../types';
+import type { PerturbationFiltersSortState } from './PerturbationsFilterSortState';
 
 const usePerturbationFilterSortStore = create<PerturbationFiltersSortState>(
   (set) => ({
